requirements.txt
# Google ADK and A2A Protocol
google-adk>=1.3.0
google-generativeai>=0.3.0

# FastAPI for A2A HTTP endpoints
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# External API integrations
requests>=2.31.0
httpx>=0.25.0
firecrawl-py>=0.0.16
exa-py>=1.0.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.25.0
python-dateutil>=2.8.2

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.2.0

# Caching
diskcache>=5.6.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0